fix:rename  forder module on models
