+++ src/app.module.ts
@@ -3,23 +3,23 @@ import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersModule } from './modules/user/user.module';
-import { UnionsModule } from './modules/unions/unions.module';
-import { User } from './modules/user/entity/user.entity';
-import { Unions } from './modules/unions/entity/unions.entity';
-import { Language } from './modules/language/entity/language.entity';
-import { Workspaces } from './modules/workspaces/entity/workspaces.entity';
-import { UnionTypes } from './modules/union_types/entity/union_type.entity';
-import { Permissions } from './modules/permissions/entity/permissions.entity';
-import { Refresh_Token } from './modules/refresh_tokens/entity/refresh_tokens.entity';
-import { WorkSpaceKeys } from './modules/workspace_keys/entity/workspace_keys.entity';
-import { UnionsDataService } from './modules/unions_data/unions_data.service';
-import { UnionsDataModule } from './modules/unions_data/unions_data.module';
-import { Union_Data } from './modules/unions_data/entity/unions_data.entity';
-import { ScriptsController } from './modules/scripts/scripts.controller';
-import { ScriptsService } from './modules/scripts/scripts.service';
-import { ScriptsModule } from './modules/scripts/scripts.module';
-import { Scripts } from './modules/scripts/entity/scripts.entity';
+import { UsersModule } from './models/user/user.module';
+import { UnionsModule } from './models/unions/unions.module';
+import { User } from './models/user/entity/user.entity';
+import { Unions } from './models/unions/entity/unions.entity';
+import { Language } from './models/language/entity/language.entity';
+import { Workspaces } from './models/workspaces/entity/workspaces.entity';
+import { UnionTypes } from './models/union_types/entity/union_type.entity';
+import { Permissions } from './models/permissions/entity/permissions.entity';
+import { Refresh_Token } from './models/refresh_tokens/entity/refresh_tokens.entity';
+import { WorkSpaceKeys } from './models/workspace_keys/entity/workspace_keys.entity';
+import { UnionsDataService } from './models/unions_data/unions_data.service';
+import { UnionsDataModule } from './models/unions_data/unions_data.module';
+import { Union_Data } from './models/unions_data/entity/unions_data.entity';
+import { ScriptsController } from './models/scripts/scripts.controller';
+import { ScriptsService } from './models/scripts/scripts.service';
+import { ScriptsModule } from './models/scripts/scripts.module';
+import { Scripts } from './models/scripts/entity/scripts.entity';
 @Module({
   imports: [
     ConfigModule.forRoot({
