+++ src/models/union_types/entity/union_type.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, Column, PrimaryGeneratedColumn, OneToMany } from 'typeorm';
-import { Unions } from 'src/modules/unions/entity/unions.entity';
+import { Unions } from 'src/models/unions/entity/unions.entity';
 @Entity()
 export class UnionTypes {
   @PrimaryGeneratedColumn()
